% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abilities.R
\name{abilities}
\alias{abilities}
\title{Extract ability parameter estimates from a cirtfit object}
\usage{
abilities(x, type = "EAP")
}
\arguments{
\item{x}{A cirtfit object.}

\item{type}{Defaults to "EAP" which is the only currently available option.}
}
\description{
Extract ability parameter estimates from a cirtfit object. There are different possibiliries for ability parameter estimation; currently, the only available option is EAP estimates.
}
