% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_model_fitting.R
\name{add_inference}
\alias{add_inference}
\title{Function that computes standard errors, z- and p-values for Wald tests, and CI.}
\usage{
add_inference(model, prob = 0.95)
}
\arguments{
\item{model}{A cirt model. As returned by the cirt function.}

\item{prob}{Probability for the CI. Defaults to 0.95}
}
\description{
Function that computes standard errors, z- and p-values for Wald tests, and CI.
}
