% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{anova.cirtfit}
\alias{anova.cirtfit}
\title{Compare two nested cirtfit objects}
\usage{
\method{anova}{cirtfit}(object, ...)
}
\arguments{
\item{object}{A cirtfit object.}

\item{...}{Additional cirtfit objects.}
}
\description{
Compare two nested cirtfit objects in using a likelihood ratio test, the AIC and the BIC.
}
\details{
The degrees of freedom used for the AIC and BIC are the number of model parameters. For the LR test, the degrees of freedom are the difference in the number of model parameters between the two models.
}
