% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirt_mice.R
\name{cirt_mice}
\alias{cirt_mice}
\title{Model fitting function for count data IRT models using \link[countirt]{cirt} on mids objects (as returned after performing multiple imputations with \link[mice]{mice}).}
\usage{
cirt_mice(
  model,
  mids_object,
  family,
  item_offset = NULL,
  person_offset = NULL,
  data_long = FALSE,
  person_id = NULL,
  control = list(n_nodes = 121, thres = Inf, prob = 0, init_disp_one = TRUE, maxiter =
    1000, convtol = 1e-05, ctol_maxstep = 1e-08, m_method = "nleqslv", convcrit =
    "marglik")
)
}
\arguments{
\item{model}{A string specifying the model. See \link[countirt]{cirt} for more information.}

\item{mids_object}{A mids object as obtained after performing multiple imputations on a dataset using \link[mice]{mice}.}

\item{family}{A string indicating the count data family, can be either "cmp" or "poisson".}

\item{item_offset}{Either a scalar (for same offset for all items) or a vector of the same length as the number of items with an offset to be added to the prediction term (on log scale).}

\item{person_offset}{Either a scalar (for same offset for all persons) or a vector of the same length as the number of persons with an offset to be added to the prediction term (on log scale).}

\item{data_long}{A boolean. Indicates whether data is in long format. If FALSE, expects data in wide format. Defaults to FALSE.}

\item{person_id}{A character string. Name of the column with person id in long format data frame. Only necessary if data_long = TRUE.}

\item{control}{A list providing control parameters for \link[countirt]{cirt}.}
}
\description{
Model fitting function for count data IRT models using \link[countirt]{cirt} on mids objects (as returned after performing multiple imputations with \link[mice]{mice}).
}
