% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{item_curves}
\alias{item_curves}
\title{Function to plot item response curves for cirtfit object}
\usage{
item_curves(x, shading = TRUE, grid = FALSE, theta_grid = 300)
}
\arguments{
\item{x}{A cirtfit object.}

\item{shading}{Defaults to TRUE. Determines whether model implied SD of conditional response distributions should be drawn into the plot as shaded band (+/- 1 SD).}

\item{grid}{Defaults to TRUE. Deermines whether all item curves should be shown in one panel or each in their own panel.}

\item{theta_grid}{Defaults to 300. The number of theta values for theta grid for x axis of plot. The finer the grid, the smoother the item curve.}
}
\description{
Function to plot item response curves for cirtfit object
}
