% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_model_fitting.R
\name{mcirt_control}
\alias{mcirt_control}
\title{Control parameters for mcirt_explore, mcirt_tune_lasso, and mcirt_tune_ridge.}
\usage{
mcirt_control(
  start_values = NULL,
  em_type = "gh",
  n_nodes = 12,
  n_samples = 3000,
  fcov_prior = NULL,
  truncate_grid = TRUE,
  maxiter = 1000,
  convtol = 1e-05,
  n_samples_conv = 20,
  final_n_samples = 6000,
  ctol_maxstep = 1e-08,
  ctol_lasso = 0.001,
  m_method = "nleqslv",
  convcrit = "marglik"
)
}
\arguments{
\item{start_values}{Optional external start values, must be of the same structure as mcirt start values (i.e., first all alphas for one trait, then all alphas for the next, etc., then all deltas, then all log nus). Dispersions must be provided on log scale. Parameters must be named as expected bei mcirt, that is of the form: "alphaj_thetal" with j in {1, ..., M} and l in {1, ..., L}, where M is the number of items and L is the number of traits (factors); "deltaj" with j in {1, ..., M}, and "log_dispj" with j in {1, ..., M}.}

\item{em_type}{A string. Either "gh" or "mc". Specifies whether Gauss-Hermite Quadrature or Monte Carlo integration is used to approximate integrals in EM algorithm.}

\item{n_nodes}{An integer. Only used in conjunction with `em_type = "gh"`. Specifies number of nodes per trait.}

\item{n_samples}{An integer. Only used in conjunction with `em_type = "mc"`. Specifies number of MC samples.}

\item{fcov_prior}{A list. First element must be the vector for mu (of length L, which is the number of traits) and second element must be the covariance matrix sigma (of dimensionality L x L).}

\item{truncate_grid}{A boolean. Only used in conjunction with `em_type = "gh"`. Indicates whether quadrature points with very low weight are removed.}

\item{maxiter}{An integer. Maximum number of EM iterations.}

\item{convtol}{A real. Convergence tolerance for EM algorithm.}

\item{n_samples_conv}{An integer. Only used in conjunction with `em_type = "mc"`.}

\item{final_n_samples}{An integer. Only used in conjunction with `em_type = "mc"`.}

\item{ctol_maxstep}{A real. Convergence tolerance for M step of EM algorithm.}

\item{ctol_lasso}{A real. Convergence tolerance for Lasso coordinate descent algorithm in M step of EM algorithm.}

\item{m_method}{A string. Can be "nleqslv". Specifying optimizer used for M step of EM algorithm,}

\item{convcrit}{A string. Can be "marglik" or "params". Specifying whether to assess convergence on marginal likelihood or parameter estimates.}
}
\description{
Control parameters for mcirt_explore, mcirt_tune_lasso, and mcirt_tune_ridge.
}
