% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_model_fitting.R
\name{mcirt_tune_ridge}
\alias{mcirt_tune_ridge}
\title{Tune ridge regularization for exploratory multi-dimensional count data IRT models.}
\usage{
mcirt_tune_ridge(
  nfactors,
  data,
  family,
  penalize_grid,
  tuning_crit = "AIC",
  alpha_constraints = NULL,
  disp_constraints = NULL,
  control = list(start_values = NULL, em_type = "gh", n_nodes = 12, n_samples = 3000,
    truncate_grid = TRUE, maxiter = 1000, convtol = 1e-05, n_samples_conv = 20,
    final_n_samples = 6000, ctol_maxstep = 1e-08, ctol_lasso = 0.001, m_method =
    "nleqslv", convcrit = "marglik")
)
}
\arguments{
\item{nfactors}{An integer. The number of factors to be extracted.}

\item{data}{A data frame (or matrix) in wide format where each column corresponds to participants' responses to one item.}

\item{family}{A string indicating the count data family, can be either "cmp" or "poisson".}

\item{penalize_grid}{A vector of reals for possible penalization parameter values.}

\item{tuning_crit}{A string. Can be either "AIC" or "BIC". Specifies after what criterion tuning parameter should be selected.}

\item{alpha_constraints}{A matrix with constraints for the dispersion parameter matrix. The matrix should be of dimensionality: nfactor rows and ncol(data) columns. Each row represents a factor, each column an item. You can currently either specify to estimate a discrimination parameter freely (by entering `NA` into the `alpha_constraints` matrix) or fix it to a specific value (by entering that value into the `alpha_constraints` matrix) Support for further constraints  (e.g., equality constraints) may be added in the future.}

\item{disp_constraints}{A vector. Specifies constraints for the dispersion parameters (on the original scale). Defaults to no constraints. Currently, you can provide a vector of values (of length ncol(data), i.e., number of items) to which dispersions should be fixed. Support for partial fixations (and free estimation of the rest) and equality constraints may be added in}

\item{control}{A list providing control parameters for the estimation to be passed to \link[countirt]{mcirt_explore}.}
}
\description{
Tune ridge regularization for exploratory multi-dimensional count data IRT models.
}
