% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_model_fitting.R
\name{tune_ridge}
\alias{tune_ridge}
\title{Tune lasso regularization for exploratory multi-dimensional count data IRT models.}
\usage{
tune_ridge(x)
}
\arguments{
\item{nfactors}{An integer. The number of factors to be extracted.}

\item{data}{A data frame (or matrix) in wide format where each column corresponds to participants' responses to one item.}

\item{family}{A string indicating the count data family, can be either "cmp" or "poisson".}

\item{penalize_grid}{A vector of reals for possible penalization parameter values.}

\item{tuning_crit}{A string. Can be either "AIC" or "BIC". Specifies after what criterion tuning parameter should be selected.}

\item{alpha_constraints}{A matrix with constraints for the dispersion parameter matrix. The matrix should be of dimensionality: nfactor rows and ncol(data) columns. Each row represents a factor, each column an item. You can currently either specify to estimate a discrimination parameter freely (by entering `NA` into the `alpha_constraints` matrix) or fix it to a specific value (b< entering that value into the `alpha_constraints` matrix) Support for further constraints  (e.g., equality constraints) may be added in the future.}

\item{control}{A list providing control parameters for the estimation to be passed to \link[countirt]{mcirt_explore}.}
}
\description{
Tune lasso regularization for exploratory multi-dimensional count data IRT models.
}
