% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_model_fitting.R
\name{cirt}
\alias{cirt}
\title{model fitting function for count data irt models}
\usage{
cirt(
  data,
  family,
  n_nodes = 121,
  stand_errors = FALSE,
  constraints = list(fix_disps = NULL, fix_alphas = NULL, same_disps = FALSE,
    same_alphas = FALSE),
  control = list(thres = Inf, prob = 0, init_disp_one = TRUE, maxiter = 1000, convtol =
    1e-05, ctol_maxstep = 1e-08, m_method = "nleqslv", convcrit = "marglik")
)
}
\arguments{
\item{data}{data matrix, each column must correspond to one item, each row to an observational unit (usually a person)}

\item{family}{a string indicating the count data family, can be either "cmp" or "poisson"}

\item{n_nodes}{integer, number of quadrtaure nodes, defaults to 121, for the 2CMP model, no less than 100 quadrature nodes are recommended}

\item{stand_errors}{boolean, indicates whether standard errors for model parameters should be estimated, defaults to FALSE}

\item{constraints}{a list, indicating the constraints for the model, note that at the current moment, the possible constraints are mutually exclusive (this functionality will be extended in the future).}

\item{control}{a list, providing control parameters for the estimation}
}
\description{
model fitting function for count data irt models
}
